Extract the store from index to its own module
