+++ src/store/root-reducer.js
@@ -1,6 +1,6 @@
-import authReducer from '../slices/auth/reducer';
-import adminReducer from '../slices/admin/reducer';
-import themeReducer from '../slices/theme/reducer';
+import authReducer from './slices/auth/reducer';
+import adminReducer from './slices/admin/reducer';
+import themeReducer from './slices/theme/reducer';
 import { firebaseReducer } from 'react-redux-firebase';
 
 import { combineReducers } from 'redux';
